@@ -11,31 +11,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.4",
-    "@angular/cdk": "^7.2.1",
-    "@angular/common": "^7.1.4",
-    "@angular/compiler": "^7.1.4",
-    "@angular/core": "^7.1.4",
-    "@angular/forms": "^7.1.4",
-    "@angular/http": "^7.1.4",
-    "@angular/material": "^7.2.1",
-    "@angular/material-moment-adapter": "^7.2.1",
-    "@angular/platform-browser": "^7.1.4",
-    "@angular/platform-browser-dynamic": "^7.1.4",
-    "@angular/router": "^7.1.4",
-    "core-js": "^2.6.1",
+    "@angular/animations": "^7.2.9",
+    "@angular/cdk": "^7.3.4",
+    "@angular/common": "^7.2.9",
+    "@angular/compiler": "^7.2.9",
+    "@angular/core": "^7.2.9",
+    "@angular/forms": "^7.2.9",
+    "@angular/http": "^7.2.9",
+    "@angular/material": "^7.3.4",
+    "@angular/material-moment-adapter": "^7.3.4",
+    "@angular/platform-browser": "^7.2.9",
+    "@angular/platform-browser-dynamic": "^7.2.9",
+    "@angular/pwa": "^0.12.4",
+    "@angular/router": "^7.2.9",
+    "@angular/service-worker": "^7.2.12",
+    "core-js": "^2.6.5",
     "material-design-icons": "^3.0.1",
-    "moment": "^2.23.0",
-    "rxjs": "^6.3.3",
-    "rxjs-compat": "^6.3.3",
-    "zone.js": "^0.8.27"
+    "moment": "^2.24.0",
+    "rxjs": "^6.4.0",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12.1",
-    "@angular/cli": "^7.2.1",
-    "@angular/compiler-cli": "^7.2.0",
-    "@angular/language-service": "^7.2.0",
-    "@types/jasmine": "^2.8.14",
+    "@angular-devkit/build-angular": "^0.13.7",
+    "@angular/cli": "^7.3.6",
+    "@angular/compiler-cli": "^7.2.9",
+    "@angular/language-service": "^7.2.9",
+    "@types/jasmine": "^2.8.16",
     "@types/jasminewd2": "^2.0.6",
     "@types/node": "~8.9.4",
     "codelyzer": "~4.2.1",
@@ -43,7 +44,7 @@
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^3.1.4",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.4",
+    "karma-coverage-istanbul-reporter": "^2.0.5",
     "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "^5.4.2",
