@@ -1,12 +1,13 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { HttpClientModule }    from '@angular/common/http';
-import { HomeComponent } from './components/home/home.component';
+import { HttpClientModule } from '@angular/common/http';
+import { HomeComponent } from './home.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import {
   MatInputModule,
   MatButtonModule,
   MatIconModule,
+  MatCardModule,
 } from '@angular/material';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 
@@ -15,15 +16,15 @@ import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
     MatInputModule,
     MatButtonModule,
     MatIconModule,
+    MatCardModule,
     BrowserAnimationsModule,
-
- 	  BrowserModule,
+    BrowserModule,
     HttpClientModule,
     FormsModule,
     ReactiveFormsModule
   ],
   declarations: [
-  	HomeComponent
+    HomeComponent
   ]
 })
 export class HomeModule { }
