@@ -1,13 +1,12 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA, MatSnackBar } from '@angular/material';
 import { FormGroup, Validators, FormControl } from '@angular/forms';
-import * as moment from 'moment';
 import { ServiceService, ServiceStruct } from 'src/app/_services/service.service';
 
 @Component({
   selector: 'app-dialog-update-service',
   templateUrl: './dialog-update-service.component.html',
-  styleUrls: ['./dialog-update-service.component.css']
+  styleUrls: ['../../../_css/dialogs_form.scss']
 })
 export class DialogUpdateServiceComponent implements OnInit {
   form: FormGroup;
