@@ -14,15 +14,11 @@ import {
   MatCardModule } from '@angular/material';
 import { AvtoService } from '../_services/avto.service';
 import { SelectedAvtoModule } from '../shared/selected-avto/selected-avto.module';
-import { ItemAvtoModule } from '../shared/item-avto/item-avto.module';
 import { GroupService } from '../_services/groups.service';
-import { ItemGroupComponent } from './item-group/item-group.component';
 import { ItemServiceComponent } from './item-service/item-service.component';
 import { ServiceService } from '../_services/service.service';
-import { DialogCreateAvtoComponent } from './dialogs/dialog-create-avto/dialog-create-avto.component';
+import { DialogCreateAvtoComponent } from '../app_components/dialog-create-avto/dialog-create-avto.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { DialogCreateGroupComponent } from './dialogs/dialog-create-group/dialog-create-group.component';
-import { DialogUpdateGroupComponent } from './dialogs/dialog-update-group/dialog-update-group.component';
 import { DialogCreateServiceComponent } from './dialogs/dialog-create-service/dialog-create-service.component';
 import { DialogUpdateServiceComponent } from './dialogs/dialog-update-service/dialog-update-service.component';
 import { OrderPipeModule } from '../_pipes/order.pipe.module';
@@ -30,19 +26,15 @@ import { OrderPipeModule } from '../_pipes/order.pipe.module';
 @NgModule({
   declarations: [
     ServicesComponent,
-    ItemGroupComponent,
     ItemServiceComponent,
     DialogCreateAvtoComponent,
     DialogCreateAvtoComponent,
-    DialogCreateGroupComponent,
-    DialogUpdateGroupComponent,
     DialogCreateServiceComponent,
     DialogUpdateServiceComponent,
   ],
   imports: [
     CommonModule,
     SelectedAvtoModule,
-    ItemAvtoModule,
     MatSnackBarModule,
     MatProgressSpinnerModule,
     MatProgressBarModule,
@@ -64,8 +56,6 @@ import { OrderPipeModule } from '../_pipes/order.pipe.module';
   ],
   entryComponents: [
     DialogCreateAvtoComponent,
-    DialogCreateGroupComponent,
-    DialogUpdateGroupComponent,
     DialogCreateServiceComponent,
     DialogUpdateServiceComponent,
   ]
