@@ -1,12 +1,13 @@
 import { Component, Input } from '@angular/core';
 import { AvtoStruct } from '../../_classes/avto';
 import { MatDialog } from '@angular/material';
-import { DialogUpdateAvtoComponent } from 'src/app/services/dialogs/dialog-update-avto/dialog-update-avto.component';
+import { DialogUpdateAvtoComponent } from 'src/app/app_components/dialog-update-avto/dialog-update-avto.component';
+import { of } from 'rxjs';
 
 @Component({
   selector: 'app-selected-avto',
   templateUrl: './selected-avto.component.html',
-  styleUrls: ['./selected-avto.component.css']
+  styleUrls: ['./selected-avto.component.scss']
 })
 export class SelectedAvtoComponent {
   @Input() model: AvtoStruct;
@@ -19,7 +20,12 @@ export class SelectedAvtoComponent {
     event.preventDefault();
     this.dialog.open(DialogUpdateAvtoComponent, {
       data: this.model,
-      width: '500px'
+      width: '500px',
+      autoFocus: false,
     });
+  }
+
+  get iconURL(): string {
+    return `/api/images/small/${this.model.avto_id}.jpg?t=1`;
   }
 }
